# api/openapi.yaml
openapi: 3.0.3
info:
  title: Geek Budget - OpenAPI 3.0
  contact:
    email: ilya.korolev@outlook.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
paths:
  /v1/authorize:
    post:
      tags:
        - auth
      summary: validate user/password and return token
      security: [] # Override to indicate no security required
      operationId: authorize
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthData"
        required: true
      responses:
        "200":
          description: return token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "JWT token"
                required:
                  - token
        "401":
          description: Authentication failed
...
  /v1/matchers:
    post:
      tags:
        - matchers
      summary: create new matcher
      operationId: createMatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatcherNoID"
      responses:
        "200":
          description: created matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    get:
      tags:
        - matchers
      summary: get all matchers
      operationId: getMatchers
      responses:
        "200":
          description: matchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Matcher"
  /v1/matchers/{id}:
    get:
      tags:
        - matchers
      summary: get matcher
      operationId: getMatcher
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the matcher"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    put:
      tags:
        - matchers
      summary: update matcher
      operationId: updateMatcher
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the matcher"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatcherNoID"
      responses:
        "200":
          description: updated matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    delete:
      tags:
        - matchers
      summary: delete matcher
      operationId: deleteMatcher
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the matcher"
          required: true
          schema:
            type: "string"
            format: "uuid"
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: no body
  /v1/matchers/{id}/image:
    post:
      tags:
        - matchers
      summary: Upload matcher image
      operationId: uploadMatcherImage
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: status of the upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
    delete:
      tags:
        - matchers
      summary: delete matcher image
      operationId: deleteMatcherImage
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: updated matcher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matcher"
  /v1/matchers/check-regex:
    post:
      tags:
        - matchers
      summary: check if regex is valid and matches string (using backend's regex engine)
      operationId: checkRegex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                regex:
                  type: string
                testString:
                  type: string
              required:
                - regex
                - testString
      responses:
        "200":
          description: result of check
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:
                    type: boolean
                  isMatch:
                    type: boolean
                  error:
                    type: string
                required:
                  - isValid
                  - isMatch
        "400":
          description: bad request
  /v1/matchers/check:
    post:
      tags:
        - matchers
      summary: check if passed matcher matches given transaction
      operationId: checkMatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matcher:
                  $ref: "#/components/schemas/MatcherNoID"
                transaction:
                  $ref: "#/components/schemas/TransactionNoID"
              required:
                - matcher
                - transaction
      responses:
        "200":
          description: true/false if matcher works
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                  reason:
                    type: string
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthData:
      type: object
      properties:
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: "12345"
      required:
        - email
        - password
    Entity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      required:
        - id
...
    TransactionNoID:
      type: object
      properties:
        date:
          type: string
          format: date-time
        description:
          type: string
        place:
          type: string
        tags:
          type: array
          items:
            type: string
        partnerName:
          type: string
        partnerAccount:
          type: string
        partnerInternalId:
          type: string
          description: Internal bank's ID to be able to match later if necessary
        extra:
          type: string
          description: >-
            Stores extra data about transaction. For example could hold "variable symbol" to
            distinguish payment for the same account, but with different meaning
        unprocessedSources:
          type: string
          description: >-
            Stores FULL unprocessed transactions which was source of this transaction.
            Could be used later for detailed analysis
        externalIds:
          type: array
          items:
            type: string
          description: >-
            IDs of unprocessed transaction - to match later
        movements:
          type: array
          items:
            $ref: "#/components/schemas/Movement"
        matcherId:
          type: string
          format: uuid
          description: ID of the matcher used for this conversion (if any)
        isAuto:
          type: boolean
          description: If true, this transaction was converted automatically by the matcher
      required:
        - date
        - movements
...
    Movement:
      type: object
      properties:
        amount:
          type: number
          format: double
        currencyId:
          type: string
        accountId:
          type: string
        description:
          type: string
      required:
        - amount
        - currencyId
        # accountId is not required to create unprocessed transactions via API
        # - accountId
...
    MatcherNoID:
      type: object
      properties:
        outputDescription:
          type: string
        outputAccountId:
          type: string
          format: uuid
        outputTags:
          type: array
          items:
            type: string

        currencyRegExp:
          type: string
        partnerNameRegExp:
          type: string
        partnerAccountNumberRegExp:
          type: string
        descriptionRegExp:
          type: string
        extraRegExp:
          type: string
        confirmationHistory:
          type: array
          items:
            type: boolean
          description: >-
            List of booleans representing manual confirmations for this matcher
            (true = confirmed, false = rejected). Server enforces maximum length
            configured via application config.
        image:
          type: string
          description: "ID of the matcher image"
      required:
        - outputDescription
        - outputAccountId
    Matcher:
      type: object
      allOf:
        - $ref: "#/components/schemas/Entity"
        - $ref: "#/components/schemas/MatcherNoID"
        - type: object
          properties:
            confirmationsCount:
              type: integer
              description: >-
                Number of successful confirmations (true values) in the confirmation history.
                This shows how many times the matcher was confirmed as correct.
              example: 9
            confirmationsTotal:
              type: integer
              description: >-
                Total length of the confirmation history array.
                This is the total number of times this matcher has been evaluated.
              example: 10
          required:
            - confirmationsCount
            - confirmationsTotal
...