# backend/pkg/generated/goserver/model_matcher.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Geek Budget - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.1
 * Contact: ilya.korolev@outlook.com
 */

package goserver

type Matcher struct {
	Id string `json:"id"`

	OutputDescription string `json:"outputDescription"`

	OutputAccountId string `json:"outputAccountId"`

	OutputTags []string `json:"outputTags,omitempty"`

	CurrencyRegExp string `json:"currencyRegExp,omitempty"`

	PartnerNameRegExp string `json:"partnerNameRegExp,omitempty"`

	PartnerAccountNumberRegExp string `json:"partnerAccountNumberRegExp,omitempty"`

	DescriptionRegExp string `json:"descriptionRegExp,omitempty"`

	ExtraRegExp string `json:"extraRegExp,omitempty"`

	// List of booleans representing manual confirmations for this matcher (true = confirmed, false = rejected). Server enforces maximum length configured via application config.
	ConfirmationHistory []bool `json:"confirmationHistory,omitempty"`

	// ID of the matcher image
	Image string `json:"image,omitempty"`

	// Number of successful confirmations (true values) in the confirmation history. This shows how many times the matcher was confirmed as correct.
	ConfirmationsCount int32 `json:"confirmationsCount"`

	// Total length of the confirmation history array. This is the total number of times this matcher has been evaluated.
	ConfirmationsTotal int32 `json:"confirmationsTotal"`
}

type MatcherInterface interface {
	GetId() string
	GetOutputDescription() string
	GetOutputAccountId() string
	GetOutputTags() []string
	GetCurrencyRegExp() string
	GetPartnerNameRegExp() string
	GetPartnerAccountNumberRegExp() string
	GetDescriptionRegExp() string
	GetExtraRegExp() string
	GetConfirmationHistory() []bool
	GetImage() string
	GetConfirmationsCount() int32
	GetConfirmationsTotal() int32
}

func (c *Matcher) GetId() string {
	return c.Id
}
func (c *Matcher) GetOutputDescription() string {
	return c.OutputDescription
}
func (c *Matcher) GetOutputAccountId() string {
	return c.OutputAccountId
}
func (c *Matcher) GetOutputTags() []string {
	return c.OutputTags
}
func (c *Matcher) GetCurrencyRegExp() string {
	return c.CurrencyRegExp
}
func (c *Matcher) GetPartnerNameRegExp() string {
	return c.PartnerNameRegExp
}
func (c *Matcher) GetPartnerAccountNumberRegExp() string {
	return c.PartnerAccountNumberRegExp
}
func (c *Matcher) GetDescriptionRegExp() string {
	return c.DescriptionRegExp
}
func (c *Matcher) GetExtraRegExp() string {
	return c.ExtraRegExp
}
func (c *Matcher) GetConfirmationHistory() []bool {
	return c.ConfirmationHistory
}
func (c *Matcher) GetImage() string {
	return c.Image
}
func (c *Matcher) GetConfirmationsCount() int32 {
	return c.ConfirmationsCount
}
func (c *Matcher) GetConfirmationsTotal() int32 {
	return c.ConfirmationsTotal
}

// AssertMatcherRequired checks if the required fields are not zero-ed
func AssertMatcherRequired(obj Matcher) error {
	elements := map[string]interface{}{
		"id":                 obj.Id,
		"outputDescription":  obj.OutputDescription,
		"outputAccountId":    obj.OutputAccountId,
		"confirmationsCount": obj.ConfirmationsCount,
		"confirmationsTotal": obj.ConfirmationsTotal,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertMatcherConstraints checks if the values respects the defined constraints
func AssertMatcherConstraints(obj Matcher) error {
	return nil
}